requirements.txt
amqp==2.6.0
asgiref==3.2.10
astroid==2.4.2
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-background-tasks==1.2.5
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-compat==1.0.15
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-feedparser==0.2.1
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
feedparser==5.2.1
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
pylint-django==2.1.0
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
